Renamed files and qfixed mypy
